requirements.txt
aiogram==3.10.0
fastapi==0.111.0
uvicorn==0.29.0
python-dotenv==1.0.1
jinja2==3.1.4
asyncpg==0.29.0
pydantic==2.7.1
python-multipart==0.0.9